+++ sdk/hardwaresecuritymodules/azure-mgmt-hardwaresecuritymodules/azure/mgmt/hardwaresecuritymodules/aio/_client.py
@@ -1,3 +1,4 @@
+# pylint: disable=line-too-long,useless-suppression
 # coding=utf-8
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
@@ -7,8 +8,8 @@
 # --------------------------------------------------------------------------
 
 from copy import deepcopy
+import sys
 from typing import Any, Awaitable, Optional, TYPE_CHECKING, cast
-from typing_extensions import Self
 
 from azure.core.pipeline import policies
 from azure.core.rest import AsyncHttpResponse, HttpRequest
@@ -27,14 +28,23 @@ from .operations import (
     CloudHsmClustersOperations,
     DedicatedHsmOperations,
     Operations,
+    PaymentHsmClusterPrivateEndpointConnectionsOperations,
+    PaymentHsmClusterPrivateLinkResourcesOperations,
+    PaymentHsmClustersOperations,
     PrivateEndpointConnectionsOperations,
 )
 
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    from typing_extensions import Self  # type: ignore
+
 if TYPE_CHECKING:
+    from azure.core import AzureClouds
     from azure.core.credentials_async import AsyncTokenCredential
 
 
-class HardwareSecurityModulesMgmtClient:  # pylint: disable=too-many-instance-attributes
+class HardwareSecurityModulesMgmtClient:  # pylint: disable=too-many-instance-attributes,docstring-keyword-should-match-keyword-only
     """The Azure management API provides a RESTful set of web services that interact with Hardware
     Security Modules.
 
@@ -46,6 +56,13 @@ class HardwareSecurityModulesMgmtClient:  # pylint: disable=too-many-instance-at
     :ivar private_endpoint_connections: PrivateEndpointConnectionsOperations operations
     :vartype private_endpoint_connections:
      azure.mgmt.hardwaresecuritymodules.aio.operations.PrivateEndpointConnectionsOperations
+    :ivar payment_hsm_clusters: PaymentHsmClustersOperations operations
+    :vartype payment_hsm_clusters:
+     azure.mgmt.hardwaresecuritymodules.aio.operations.PaymentHsmClustersOperations
+    :ivar payment_hsm_cluster_private_endpoint_connections:
+     PaymentHsmClusterPrivateEndpointConnectionsOperations operations
+    :vartype payment_hsm_cluster_private_endpoint_connections:
+     azure.mgmt.hardwaresecuritymodules.aio.operations.PaymentHsmClusterPrivateEndpointConnectionsOperations
     :ivar cloud_hsm_cluster_private_link_resources: CloudHsmClusterPrivateLinkResourcesOperations
      operations
     :vartype cloud_hsm_cluster_private_link_resources:
@@ -63,24 +80,39 @@ class HardwareSecurityModulesMgmtClient:  # pylint: disable=too-many-instance-at
     :ivar dedicated_hsm: DedicatedHsmOperations operations
     :vartype dedicated_hsm:
      azure.mgmt.hardwaresecuritymodules.aio.operations.DedicatedHsmOperations
+    :ivar payment_hsm_cluster_private_link_resources:
+     PaymentHsmClusterPrivateLinkResourcesOperations operations
+    :vartype payment_hsm_cluster_private_link_resources:
+     azure.mgmt.hardwaresecuritymodules.aio.operations.PaymentHsmClusterPrivateLinkResourcesOperations
     :param credential: Credential used to authenticate requests to the service. Required.
     :type credential: ~azure.core.credentials_async.AsyncTokenCredential
     :param subscription_id: The ID of the target subscription. The value must be an UUID. Required.
     :type subscription_id: str
     :param base_url: Service host. Default value is None.
     :type base_url: str
-    :keyword api_version: The API version to use for this operation. Default value is "2025-03-31".
-     Note that overriding this default value may result in unsupported behavior.
+    :keyword cloud_setting: The cloud setting for which to get the ARM endpoint. Default value is
+     None.
+    :paramtype cloud_setting: ~azure.core.AzureClouds
+    :keyword api_version: The API version to use for this operation. Known values are
+     "2025-12-01-preview" and None. Default value is None. If not set, the operation's default API
+     version will be used. Note that overriding this default value may result in unsupported
+     behavior.
     :paramtype api_version: str
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
      Retry-After header is present.
     """
 
     def __init__(
-        self, credential: "AsyncTokenCredential", subscription_id: str, base_url: Optional[str] = None, **kwargs: Any
+        self,
+        credential: "AsyncTokenCredential",
+        subscription_id: str,
+        base_url: Optional[str] = None,
+        *,
+        cloud_setting: Optional["AzureClouds"] = None,
+        **kwargs: Any
     ) -> None:
         _endpoint = "{endpoint}"
-        _cloud = kwargs.pop("cloud_setting", None) or settings.current.azure_cloud  # type: ignore
+        _cloud = cloud_setting or settings.current.azure_cloud  # type: ignore
         _endpoints = get_arm_endpoints(_cloud)
         if not base_url:
             base_url = _endpoints["resource_manager"]
@@ -89,6 +121,7 @@ class HardwareSecurityModulesMgmtClient:  # pylint: disable=too-many-instance-at
             credential=credential,
             subscription_id=subscription_id,
             base_url=cast(str, base_url),
+            cloud_setting=cloud_setting,
             credential_scopes=credential_scopes,
             **kwargs
         )
@@ -125,6 +158,12 @@ class HardwareSecurityModulesMgmtClient:  # pylint: disable=too-many-instance-at
         self.private_endpoint_connections = PrivateEndpointConnectionsOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
+        self.payment_hsm_clusters = PaymentHsmClustersOperations(
+            self._client, self._config, self._serialize, self._deserialize
+        )
+        self.payment_hsm_cluster_private_endpoint_connections = PaymentHsmClusterPrivateEndpointConnectionsOperations(
+            self._client, self._config, self._serialize, self._deserialize
+        )
         self.cloud_hsm_cluster_private_link_resources = CloudHsmClusterPrivateLinkResourcesOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
@@ -138,6 +177,9 @@ class HardwareSecurityModulesMgmtClient:  # pylint: disable=too-many-instance-at
             self._client, self._config, self._serialize, self._deserialize
         )
         self.dedicated_hsm = DedicatedHsmOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.payment_hsm_cluster_private_link_resources = PaymentHsmClusterPrivateLinkResourcesOperations(
+            self._client, self._config, self._serialize, self._deserialize
+        )
 
     def send_request(
         self, request: HttpRequest, *, stream: bool = False, **kwargs: Any